fix: incorrect set controller
